FIX separate declaration and definition in order to allow multiple include and avoid linker errors
